feat(createSvgIcon): initial implementation
- deprecate unstable_createSvgIcon in favor of createSvgIcon_unstable
- new name for old util, no changes
