+++ libs/spark/src/createSvgIcon_unstable/index.ts
@@ -1,0 +1,1 @@
+export { default } from './createSvgIcon_unstable';
